moved user functionality to blueprint for modularity
